# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CKB Agent Wallet Library API — programmable wallet for AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all agent wallets
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentWallet"
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Create a new agent wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletInput"
      responses:
        "201":
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWallet"
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get a wallet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWallet"
        "404":
          description: Wallet not found
    delete:
      operationId: deleteWallet
      tags: [wallets]
      summary: Delete a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /wallets/{id}/balance:
    get:
      operationId: getWalletBalance
      tags: [wallets]
      summary: Get wallet balance from CKB node
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
        "404":
          description: Wallet not found

  # ── Cells ─────────────────────────────────────────────────────────────────
...
  /wallets/{id}/kill:
    post:
      operationId: activateKillSwitch
      tags: [safety]
      summary: Activate kill switch — disables all agent operations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kill switch activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWallet"
        "404":
          description: Wallet not found
  /wallets/{id}/restore:
    post:
      operationId: restoreWallet
      tags: [safety]
      summary: Restore wallet after kill switch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Wallet restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentWallet"
        "404":
          description: Wallet not found

  # ── Fiber Channels ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AgentWallet:
      type: object
      required: [id, name, address, publicKey, network, status, isKilled, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          description: CKB address (secp256k1-based)
        publicKey:
          type: string
          description: Hex-encoded compressed public key
        network:
          type: string
          enum: [mainnet, testnet]
        status:
          type: string
          enum: [active, killed]
        isKilled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WalletInput:
      type: object
      required: [name, network]
      properties:
        name:
          type: string
          minLength: 1
        network:
          type: string
          enum: [mainnet, testnet]
        passphrase:
          type: string
          description: Passphrase for encrypting the private key. If omitted, uses server default.
    WalletBalance:
      type: object
      required: [walletId, address, totalCapacity, freeCapacity, cellCount]
      properties:
        walletId:
          type: integer
        address:
          type: string
        totalCapacity:
          type: string
          description: Total capacity in shannons (1 CKB = 100,000,000 shannons)
        freeCapacity:
          type: string
          description: Spendable capacity in shannons
        cellCount:
          type: integer

    # ── Cells ───────────────────────────────────────────────────────────────
...