# lib/api-zod/src/generated/types/walletBalance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CKB Agent Wallet Library API — programmable wallet for AI agents
 * OpenAPI spec version: 0.1.0
 */

export interface WalletBalance {
  walletId: number;
  address: string;
  /** Total capacity in shannons (1 CKB = 100,000,000 shannons) */
  totalCapacity: string;
  /** Spendable capacity in shannons */
  freeCapacity: string;
  cellCount: number;
}